Modifiable unresolved color (#1)
For unresolved nodes/notes, the previous colour was 
--background-modifier-error, which looks  a lot like the colour of the focused
node (--color-fill-focused). This can be confusing if you have a lot of
unresolved nodes. See for example
https://github.com/jdanielmourao/obsidian-sanctum/issues/86, you can see
in the graph view that many nodes look like they are the active one.

This PR makes the unresolved node colour modifiable in Style Settings.